Adds single core launch support for AMD
This patch enables the ability to launch the VMM on AMD on a
single core. It uses an old hack to deal with a lack of our
own CR3, which will be removed once we have the CR3 part
added, and as a result, it only works on single core. It also
cannot be stopped, as that logic is really specific to the
way that the VMM was launched which will change once all of
this code is moved to C++, so that functionality will come
later as well.