Improve NMI handling
This patch modifies the way NMIs are handled when running in the
VMM. The old approach unconditionally enables NMI-window exiting
and injects the NMI when the window exit occurs. This works fine
as long the vcpu that caused the exit is the same one that is
subsequently entered.

However if a different one is entered, since NMIs can fire on any
instruction in the VMM, it is possible for the NMI-window exit to be
enabled on the vcpu that is *not* being entered. In this case the NMI
sits in the non-current VMCS for an indeterminate amount of time.

These changes ensure that no matter when the NMI is taken, the NMI-window
exit is always enabled on the VMCS that is being entered (i.e. with
vmlaunch or vmresume). It also eliminates races with accesses to the
VMCS field from C++ that control the window exit, by ensuring that
the field is only read and written in the vmcs_resume or
vmcs_launch code.