First VMExit Failure Logic
This patch cleans up the first VMExit failure logic so that
it actually makes sense and works as intended. Now, if you
have an exception, or error on your first exit, the hypervior
will gracefully exit instead of halting, or pretending to
gracefully exit when in fact it was just getting luck.