Loader: Build the arch target defined by CMake
This patch fixes an issue that could arise if the following conditions
are met:
- set HYPERVISOR_TARGET_ARCH to Intel in CMake
- build the loader for Linux on an AMD machine
- On AMD, `make driver_quick`, then `make start` -> computer hangs

This occurs because the loader is built using the architecture of the
build machine while the hypervisor is built using the target arch set
during cmake configuration which can be different. The cpu checks are
done inside the Linux driver which will succeed and load the hypervisor
which is using Intel specific instructions causing the machine to hang
when it tries to start on AMD.

We now use the cmake HYPERVISOR_TARGET_ARCH variable in Makefile which
will keep the hypervisor and the Linux driver in sync.